Renamed navigator to show use
